refactor(model): 拆分设置模型与固定数据
