+++ src-tauri/src/hardcode_data/dev_samples.rs
@@ -1,0 +1,8 @@
+//! 测试和示例用的固定数据。
+
+#[allow(dead_code)] // 测试调用
+pub const SAMPLE_SERVER_CORE_LFS_URL: &str = "https://cnb.cool/SeaLantern-studio/ServerCore-Mirror/-/lfs/7f717a1fe4e30ee53671540f09142808efced1ef19f5d68219afa458e048ebf5?name=arclight-fabric-1.20.4-1.0.4-80ec5df.jar";
+
+#[allow(dead_code)] // 测试调用
+pub const SAMPLE_DOWNLOAD_MANAGER_URL: &str =
+    "https://files.mcjars.app/mohist/1.12.2/1.12.2-17e3fd09/server.jar";
