+++ src-tauri/src/hardcode_data/server_downloads.rs
@@ -1,0 +1,6 @@
+//! 服务端下载相关的固定内容。
+
+pub const SERVER_CORE_MIRROR_LINK_LIST_URL: &str =
+    "https://cnb.cool/SeaLantern-studio/ServerCore-Mirror/-/releases/download/26.02.27/jar_lfs_links.json";
+
+pub const STARTER_INSTALLER_LINKS_FILE_NAME: &str = "jar_lfs_links.json";
