+++ src-tauri/src/hardcode_data/update_sources.rs
@@ -1,0 +1,14 @@
+//! 更新源相关的固定内容。
+
+pub const CNB_BASE_URL: &str = "https://cnb.cool";
+pub const CNB_RELEASES_URL: &str = "https://cnb.cool/_next/data/f79eab91170dd314/zh/SeaLantern-studio/SeaLantern/-/releases.json?slug=SeaLantern-studio&slug=SeaLantern&slug=-&slug=releases";
+
+#[allow(dead_code)] // 发布调用
+pub const UPDATE_GITHUB_OWNER: &str = "SeaLantern-Studio";
+#[allow(dead_code)] // 发布调用
+pub const UPDATE_GITHUB_REPO: &str = "SeaLantern";
+#[allow(dead_code)] // 发布调用
+pub const UPDATE_GITHUB_API_BASE: &str = "https://api.github.com/repos";
+
+pub const UPDATE_HTTP_USER_AGENT: &str =
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36";
